requirements.txt
meshtastic
pyzmail36
python-dotenv
pypubsub
